# app/mgmt/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "此文档由gin-swagger自动生成",
        "title": "xxim-server HTTP API 文档",
        "contact": {
            "name": "showurl"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:6799",
    "basePath": "/api",
    "paths": {
...
        "/ms/add/emoji": {
            "post": {
                "description": "使用此接口新增app表情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情管理"
                ],
                "summary": "新增app表情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.AddAppMgmtEmojiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.AddAppMgmtEmojiResp"
                        }
                    }
                }
            }
        },
...
        "/ms/delete/emoji": {
            "post": {
                "description": "使用此接口删除app表情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情管理"
                ],
                "summary": "删除app表情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtEmojiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtEmojiResp"
                        }
                    }
                }
            }
        },
        "/ms/delete/emojigroup": {
            "post": {
                "description": "使用此接口删除app表情组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情组管理"
                ],
                "summary": "删除app表情组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtEmojiGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtEmojiGroupResp"
                        }
                    }
                }
            }
        },
...
        "/ms/get/emoji/detail": {
            "post": {
                "description": "使用此接口获取app表情详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情管理"
                ],
                "summary": "获取app表情详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtEmojiDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtEmojiDetailResp"
                        }
                    }
                }
            }
        },
        "/ms/get/emoji/list/all": {
            "post": {
                "description": "使用此接口获取全部app表情列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情管理"
                ],
                "summary": "获取全部app表情列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtEmojiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtEmojiResp"
                        }
                    }
                }
            }
        },
        "/ms/get/emojigroup/detail": {
            "post": {
                "description": "使用此接口获取app表情组详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情组管理"
                ],
                "summary": "获取app表情组详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtEmojiGroupDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtEmojiGroupDetailResp"
                        }
                    }
                }
            }
        },
        "/ms/get/emojigroup/list/all": {
            "post": {
                "description": "使用此接口获取全部app表情组列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情组管理"
                ],
                "summary": "获取全部app表情组列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtEmojiGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtEmojiGroupResp"
                        }
                    }
                }
            }
        },
...
        "/ms/update/emoji": {
            "post": {
                "description": "使用此接口更新app表情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情管理"
                ],
                "summary": "更新app表情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtEmojiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtEmojiResp"
                        }
                    }
                }
            }
        },
        "/ms/update/emojigroup": {
            "post": {
                "description": "使用此接口更新app表情组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app表情组管理"
                ],
                "summary": "更新app表情组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtEmojiGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtEmojiGroupResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.AddAppMgmtEmojiReq": {
            "type": "object",
            "properties": {
                "appMgmtEmoji": {
                    "$ref": "#/definitions/pb.AppMgmtEmoji"
                },
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                }
            }
        },
        "pb.AddAppMgmtEmojiResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.AppMgmtEmoji": {
            "type": "object",
            "properties": {
                "animatedUrl": {
                    "type": "string"
                },
                "cover": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdAtStr": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "staticUrl": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pb.AppMgmtEmojiGroup": {
            "type": "object",
            "properties": {
                "cover": {
                    "$ref": "#/definitions/pb.AppMgmtEmoji"
                },
                "coverId": {
                    "type": "string"
                },
                "createTime": {
                    "type": "integer"
                },
                "createTimeStr": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pb.CommonReq": {
            "type": "object",
            "properties": {
                "appVersion": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deviceId": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/pb.CommonResp_Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp_Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                7,
                8,
                9,
                10
            ],
            "x-enum-comments": {
                "CommonResp_AlertError": "alert 错误 只有一个确定按钮",
                "CommonResp_AuthError": "鉴权错误 // 应该退出登录",
                "CommonResp_ForbiddenError": "无权限",
                "CommonResp_InternalError": "内部错误",
                "CommonResp_NeedSecondPasswordError": "需要输入二级密码/设置二级密码",
                "CommonResp_RequestError": "请求错误",
                "CommonResp_RetryError": "alert 错误 有一个取消按钮 和一个重试按钮",
                "CommonResp_ToastError": "toast 错误 只有 message",
                "CommonResp_UnknownError": "未知 error"
            },
            "x-enum-varnames": [
                "CommonResp_Success",
                "CommonResp_UnknownError",
                "CommonResp_InternalError",
                "CommonResp_RequestError",
                "CommonResp_AuthError",
                "CommonResp_ToastError",
                "CommonResp_AlertError",
                "CommonResp_RetryError",
                "CommonResp_ForbiddenError",
                "CommonResp_NeedSecondPasswordError"
            ]
        },
...
        "pb.DeleteAppMgmtEmojiGroupReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.DeleteAppMgmtEmojiGroupResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
        "pb.DeleteAppMgmtEmojiReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.DeleteAppMgmtEmojiResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.GetAllAppMgmtEmojiGroupReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
        "pb.GetAllAppMgmtEmojiGroupResp": {
            "type": "object",
            "properties": {
                "appMgmtEmojiGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.AppMgmtEmojiGroup"
                    }
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "pb.GetAllAppMgmtEmojiReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
        "pb.GetAllAppMgmtEmojiResp": {
            "type": "object",
            "properties": {
                "appMgmtEmojis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.AppMgmtEmoji"
                    }
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "pb.GetAppMgmtEmojiDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pb.GetAppMgmtEmojiDetailResp": {
            "type": "object",
            "properties": {
                "appMgmtEmoji": {
                    "$ref": "#/definitions/pb.AppMgmtEmoji"
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
        "pb.GetAppMgmtEmojiGroupDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "pb.GetAppMgmtEmojiGroupDetailResp": {
            "type": "object",
            "properties": {
                "appMgmtEmojiGroup": {
                    "$ref": "#/definitions/pb.AppMgmtEmojiGroup"
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.Page": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UpdateAppMgmtEmojiGroupReq": {
            "type": "object",
            "properties": {
                "appMgmtEmojiGroup": {
                    "$ref": "#/definitions/pb.AppMgmtEmojiGroup"
                },
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                }
            }
        },
        "pb.UpdateAppMgmtEmojiGroupResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
        "pb.UpdateAppMgmtEmojiReq": {
            "type": "object",
            "properties": {
                "appMgmtEmoji": {
                    "$ref": "#/definitions/pb.AppMgmtEmoji"
                },
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                }
            }
        },
        "pb.UpdateAppMgmtEmojiResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
    }
}